Added updates to PID_Controller. Autonomous Service and more bug fixes.

+++ ros_bbauv2012/Controller/PID_Controller/include/PID_Controller/PID.h
@@ -30,7 +30,7 @@ private:
 	double Ti;
 	double Td;
 	double Imax;
-	int N;
+	double N;
 	int actMax;
 	int actMin;
 	//Controller variables
